+++ src/rules/prefer-constant.js
@@ -10,6 +10,7 @@
 module.exports = function (context) {
     const astUtil = require('../util/astUtil')
     const shouldCheckArrowFunctions = context.options[0] !== undefined ? context.options[0] : true
+    const shouldCheckFunctionDeclarations = context.options[1] !== undefined ? context.options[1] : false
 
     function isCompletelyLiteral(node) {
         switch (node.type) {
@@ -26,15 +27,28 @@ module.exports = function (context) {
         }
     }
 
-    function handleFunctionExpression(node) {
-        const valueReturnedInFirstLine = astUtil.getValueReturnedInFirstLine(node)
+    function reportIfLikeConstant(func, node) {
+        const valueReturnedInFirstLine = func(node)
         if (valueReturnedInFirstLine && isCompletelyLiteral(valueReturnedInFirstLine)) {
             context.report(node, 'Prefer _.constant over a function returning a literal')
         }
     }
 
+    function handleFunctionExpression(node) {
+        reportIfLikeConstant(astUtil.getValueReturnedInFirstLine, node)
+    }
+
+    function handleFunctionDeclaration(node) {
+        reportIfLikeConstant(astUtil.getValueReturnedInFirstLineInFuncDecl, node)
+    }
+
     return {
         FunctionExpression: handleFunctionExpression,
+        FunctionDeclaration(node) {
+            if (shouldCheckFunctionDeclarations) {
+                handleFunctionDeclaration(node)
+            }
+        },
         ArrowFunctionExpression(node) {
             if (shouldCheckArrowFunctions) {
                 handleFunctionExpression(node)
@@ -44,6 +58,9 @@ module.exports = function (context) {
 }
 
 module.exports.schema = [
+    {
+        type: 'boolean'
+    },
     {
         type: 'boolean'
     }